+++ CHANGES.md
@@ -1,10 +1,11 @@
 CHANGELOG
 =========
 
-0.1.3.dev0     (XXXX-XX-XX)
+0.1.3          (2020-11-16)
 ---------------------------
 
 * Tiles is not required anymore for add / edit mapbox base layer
+
 
 0.1.2          (2020-11-03)
 ---------------------------